Generates Frequency Statistics

Here is a sample function.

proc_freq(
  data,
  by = NULL,
  tables = NULL,
  table_options = NULL,
  weight = NULL,
  weight_options = NULL,
  view = TRUE,
  output = NULL,
  report_type = NULL,
  report_location = NULL,
  titles = NULL
)

Arguments

data
The input data frame to perform frequency calculations on.

by
An optional by group.

tables
The variable or variables to perform frequency counts on.

table_options
The option specifications for the table parameter.

weight
An optional weight parameter.

weight_options
The option specifications for the weight parameter.

view
Whether to display procedure results in the viewer. Valid values are TRUE and FALSE. Default is TRUE.

output
The directory path or libname to output data generated by the proc_freq function.

report_type
The output type for any report output. Valid values are 'HTML', 'TXT', 'PDF', 'RTF', and 'DOCX'. Multiple outputs can be requested by passing a vector of output types. If the report_location parameter is specified, the outputs will be written to that location. Otherwise, they will be written to a temp directory. The default value is NULL.

report_location
A path to write any report output requested by the print parameter. The path may be either a full path with file name, or a directory name. If the file name is not specified, the procedure type will be used. If no path is specified, files will be written to a temp directory. Default is NULL.

titles
A vector of one or more titles to use for the report output.

Value

A list of data frames that contain the requested frequency tables. The list item names are specified in the tables parameter.

Details

Some details about the sample function.