Calculates Summary Statistics

The proc_means function generates summary statistics for selected variable on the input dataset. The variables are identified on the var parameter. The statistics to perform are identified on the stats parameter. By default, results are returned as a list, of tibbles.

proc_means(
  data,
  var = NULL,
  stats = NULL,
  view = TRUE,
  report_type = NULL,
  report_location = NULL,
  report_style = NULL,
  titles = NULL,
  piped = FALSE
)

Arguments

data
The input data frame for which calculate summary statistics.

var
The variable(s) to calculate summary statistics for.

stats
A quoted vector of summary statistics keywords. Valid keywords are: "css", "cv", "n", "means", "median", "std", "min", "max", "nmiss", "nobs", "range", "sum", "var".

view
Whether to display the report in the interactive viewer. Valid values are TRUE and FALSE. Default is TRUE.

report_type
The output type for any report output. Valid values are 'HTML', 'TXT', 'PDF', 'RTF', 'DOCX'. Multiple outputs can be requested by passing a vector of output types. If the report_location parameter is specified, the outputs will be written to that location. Otherwise, they will be written to a temp directory. The default value is NULL.

report_location
A path to write any report output requested by the print parameter. The path may be either a full path with file name, or a directory name. If the file name is not specified, the procedure type will be used. If no path is specified, files will be written to a temp directory. Default is NULL.

report_style
A theme name or style object to use on the report output. See the create_style in the reporter package for additional information on styles.

titles
A vector of one or more titles to use for the report output.

piped
Whether or not the proc_freq function is part of a data pipeline. Set this parameter to TRUE if you want the function to return a single dataset instead of a list of datasets. If there is more than one table requested, the function will return the last requested table.

Value

A list of tibbles that contain the requested summary statistics. If the function is being used in a pipeline, set the piped parameter to TRUE, which will return only the last table.

Details

Some details about the sample function.

All statistics remove missing values by default. Should there be an option to include them?

Statistics Options

CSS: Corrected Sum of Squares.

CV: Coefficient of Variation.

Using Weights

Need a section explaining how to use weights.

Data Constraints

Explain limits of data for each stat option. Number of non-missing values, etc.