Transposes a Data Frame

Here is a sample function.

proc_transpose(
  data,
  by = NULL,
  var = NULL,
  id = NULL,
  idlabel = NULL,
  copy = NULL
)

Arguments

data
The input data frame for which calculate summary statistics.

by
An optional by group.

var
The variable or variables to transpose.

id
The option specifications for the table parameter.

idlabel
The variable(s) to calculate summary statistics for.

copy
A vector of variable to retain in the output data frame without transposition.

Value

A data frames that contains the transposed data.

Details

Some details about the sample function.