Calculates Summary Statistics

The proc_means function generates summary statistics for selected variables on the input dataset. The variables are identified on the var parameter. The statistics to perform are identified on the stats parameter. By default, results are displayed in the viewer interactively, and returned as a dataset or a list of datasets.

proc_means(
  data,
  by = NULL,
  class = NULL,
  var = NULL,
  stats = c("n", "mean", "std", "min", "max"),
  weight = NULL,
  titles = NULL,
  options = NULL
)

Arguments

data
The input data frame for which calculate summary statistics.

by
An optional by group. You may pass unquoted variable names to this parameter using the v function.

class
The class parameter is similar to the by parameter, but the output is different. By groups will create completely separate tables, while class groups will be continued in the same table. When a by and a class are both specified, the class will be nested in the by.

var
The variable(s) to calculate summary statistics for. You may pass unquoted variable names to this parameter using the v function.

stats
A quoted vector of summary statistics keywords. Valid keywords are: "css", "clm", "cv", "n", "lclm", "mean", "median", "mode", "min", "max", "nmiss", "nobs", "range", "std", "stderr", "sum", "uclm", "var", "q1", "q3", "p1", "p5", "p10", "p20", "p25", "p30", "p40", "p50", "p60", "p70", "p75", "p80", "p90", "p95", "p99", "qrange". You may pass unquoted variable names to this parameter using the v function. See the ProcMeansOptions page for details on the available options.

weight
An optional weight parameter.

titles
A vector of one or more titles to use for the report output.

options
An options object created with the opts function. This parameter can accept a variety of options.

Value

The requested summary statistics.

Details

Some details about the sample function.

All statistics remove missing values by default. Should there be an option to include them?

Options

CSS: Corrected Sum of Squares.

CV: Coefficient of Variation.

Using Weights

Need a section explaining how to use weights.

Data Constraints

Explain limits of data for each stat option. Number of non-missing values, etc.