Transposes a Data Frame

A function to pivot or transpose a data frame. In the default usage, the variables identified by the parameter var are transposed to become rows. The variable values in the parameter id become the new column names. The function has several more parameters to control how variables are named in the transposed data set.

proc_transpose(
  data,
  var = NULL,
  id = NULL,
  idlabel = NULL,
  copy = NULL,
  name = "NAME",
  namelabel = NULL,
  prefix = NULL,
  delimiter = ".",
  suffix = NULL
)

Arguments

data
The input data frame for which calculate summary statistics.

var
The variable or variables to transpose.

id
The variable or variables to use for the transposed column names.

idlabel
The variable to use for the transposed column labels.

copy
A vector of variables to retain in the output data without transposition.

name
Specifies the name of the variable to be used for the var values.

namelabel
The label to use for the name variable.

prefix
Contains a prefix to be used in the construction of column names.

delimiter
Specifies a delimiter to be used in the construction of column names.

suffix
Contains a suffix to be used in the construction of column names.

Value

A data frames that contains the transposed data. If a data frame is input, a data frame will be output. If a tibble is input, a tibble will be output.

Details

Some details about the sample function.

Examples

# Create data
dt <- data.frame(names = rownames(mtcars), mtcars, stringsAsFactors = FALSE)[1:5,]

# Transpose data
tdt <- proc_transpose(dt, var = c("mpg", "cyl", "disp"),
                          id = "names", name = "Variable")

# View transposed data
tdt
#   Variable Mazda RX4 Mazda RX4 Wag Datsun 710 Hornet 4 Drive Hornet Sportabout
# 1      mpg        21            21       22.8           21.4              18.7
# 2      cyl         6             6        4.0            6.0               8.0
# 3     disp       160           160      108.0          258.0             360.0